feat : 處理SDK
調整項目
1. 加入更新放款金額流程
2. 加入docker處理
3. 移除測試案例secret & key
4. 加入github action

+++ FTX.SDK/FTX.SDK/Request/PostSpotMarginOffersRequest.cs
@@ -9,10 +9,9 @@ namespace FTX.SDK.Request
     public class PostSpotMarginOffersRequest
     {
         public string coin { get; set; }
-        public double size { get; set; }
+        public float size { get; set; }
 
-
-        public double rate { get; set; }
+        public float rate { get; set; }
 
 
     }
